requirements.txt
albumentations==1.3.0
matplotlib==3.7.1
numpy==1.24.3
opencv_python_headless==4.7.0.72
Pillow==9.5.0
pycocotools==2.0.6
pytorch_lightning==2.0.4
scipy==1.10.1
shapely==2.0.1
scikit-image==0.20.0
tqdm==4.65.0
tensorboard==2.13.0